@@ -101,25 +101,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "12.2.11",
-    "@angular/common": "12.2.11",
-    "@angular/compiler": "12.2.11",
-    "@angular/core": "12.2.11",
-    "@angular/forms": "12.2.11",
-    "@angular/platform-browser": "12.2.11",
-    "@angular/platform-browser-dynamic": "12.2.11",
-    "@angular/router": "12.2.11",
-    "@angular/service-worker": "12.2.11",
+    "@angular/animations": "13.1.3",
+    "@angular/common": "13.1.3",
+    "@angular/compiler": "13.1.3",
+    "@angular/core": "13.1.3",
+    "@angular/forms": "13.1.3",
+    "@angular/platform-browser": "13.1.3",
+    "@angular/platform-browser-dynamic": "13.1.3",
+    "@angular/router": "13.1.3",
+    "@angular/service-worker": "13.1.3",
     "@ericblade/quagga2": "^1.3.1",
     "@fortawesome/angular-fontawesome": "^0.9.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.34",
     "@fortawesome/free-solid-svg-icons": "^5.15.2",
     "@microsoft/applicationinsights-web": "^2.6.2",
     "@microsoft/signalr": "^5.0.6",
-    "@ngrx/effects": "12.5.0",
-    "@ngrx/router-store": "12.5.0",
-    "@ngrx/store": "12.5.0",
-    "@nrwl/angular": "13.0.1",
+    "@ngrx/effects": "13.0.2",
+    "@ngrx/router-store": "13.0.2",
+    "@ngrx/store": "13.0.2",
+    "@nrwl/angular": "13.4.6",
     "@progress/kendo-angular-buttons": "^6.3.1",
     "@progress/kendo-angular-common": "^2.0.1",
     "@progress/kendo-angular-dateinputs": "^5.2.0",
@@ -151,26 +151,26 @@
     "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "12.2.11",
-    "@angular-eslint/eslint-plugin": "12.3.1",
-    "@angular-eslint/eslint-plugin-template": "12.3.1",
-    "@angular-eslint/template-parser": "12.3.1",
-    "@angular/cli": "12.2.11",
-    "@angular/compiler-cli": "12.2.11",
-    "@angular/language-service": "12.2.11",
-    "@angular/localize": "^12.0.4",
-    "@ngrx/schematics": "12.5.0",
-    "@ngrx/store-devtools": "12.5.0",
-    "@nrwl/cli": "13.0.1",
-    "@nrwl/eslint-plugin-nx": "13.0.1",
-    "@nrwl/jest": "13.0.1",
-    "@nrwl/tao": "13.0.1",
-    "@nrwl/workspace": "13.0.1",
+    "@angular-devkit/build-angular": "13.1.4",
+    "@angular-eslint/eslint-plugin": "13.0.1",
+    "@angular-eslint/eslint-plugin-template": "13.0.1",
+    "@angular-eslint/template-parser": "13.0.1",
+    "@angular/cli": "13.1.4",
+    "@angular/compiler-cli": "13.1.3",
+    "@angular/language-service": "13.1.3",
+    "@angular/localize": "13.1.3",
+    "@ngrx/schematics": "13.0.2",
+    "@ngrx/store-devtools": "13.0.2",
+    "@nrwl/cli": "13.4.6",
+    "@nrwl/eslint-plugin-nx": "13.4.6",
+    "@nrwl/jest": "13.4.6",
+    "@nrwl/tao": "13.4.6",
+    "@nrwl/workspace": "13.4.6",
     "@progress/kendo-licensing": "^1.2.1",
     "@types/jest": "27.0.2",
     "@types/node": "^15.3.0",
-    "@typescript-eslint/eslint-plugin": "4.32",
-    "@typescript-eslint/parser": "4.32",
+    "@typescript-eslint/eslint-plugin": "^5.10.0",
+    "@typescript-eslint/parser": "^5.10.0",
     "codelyzer": "^6.0.2",
     "cypress": "^7.3.0",
     "dotenv": "10.0.0",
@@ -180,9 +180,9 @@
     "imng-schematics": "^3.274.1",
     "jest": "27.2.3",
     "jest-junit": "^12.2.0",
-    "jest-preset-angular": "10.0.1",
+    "jest-preset-angular": "11.0.0",
     "jpeg-js": "^0.4.2",
-    "ng-packagr": "12.2.4",
+    "ng-packagr": "13.1.3",
     "openapi-ts-generator": "^3.173.5",
     "postcss": "^8.3.9",
     "postcss-import": "^14.0.2",
@@ -192,6 +192,6 @@
     "snyk": "^1.730.0",
     "ts-jest": "27.0.5",
     "ts-node": "^10.0.0",
-    "typescript": "4.3.5"
+    "typescript": "4.5.4"
   }
 }
