@@ -1,15 +1,13 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
-import { CUSTOMERS_FEATURE_KEY, State, CustomersPartialState } from './customer.reducer';
+import { CUSTOMERS_FEATURE_KEY, State } from './customer.reducer';
 import { isTruthy } from 'imng-ngrx-utils';
 
-export const getCustomersState = createFeatureSelector<CustomersPartialState, State>(
-  CUSTOMERS_FEATURE_KEY,
-);
+export const getCustomersState = createFeatureSelector<State>(CUSTOMERS_FEATURE_KEY);
 
-const getCustomers = createSelector(getCustomersState, state => state.gridData);
-const getLoading = createSelector(getCustomersState, state => state.loading);
-const getGridODataState = createSelector(getCustomersState, state => state.gridODataState);
-const getPagerSettings = createSelector(getCustomersState, state => state.gridPagerSettings);
+const getCustomers = createSelector(getCustomersState, (state) => state.gridData);
+const getLoading = createSelector(getCustomersState, (state) => state.loading);
+const getGridODataState = createSelector(getCustomersState, (state) => state.gridODataState);
+const getPagerSettings = createSelector(getCustomersState, (state) => state.gridPagerSettings);
 export const odataGridCustomerQueries = {
   getCustomers,
   getLoading,
@@ -17,17 +15,14 @@ export const odataGridCustomerQueries = {
   getPagerSettings,
 };
 
-const getCurrentCustomer = createSelector(
-  getCustomersState,
-  state => state.currentCustomer,
-);
+const getCurrentCustomer = createSelector(getCustomersState, (state) => state.currentCustomer);
 const getIsEditCustomerActive = createSelector(
   getCurrentCustomer,
-  entity => isTruthy(entity) && isTruthy(entity?.id),
+  (entity) => isTruthy(entity) && isTruthy(entity?.id),
 );
 const getIsNewCustomerActive = createSelector(
   getCurrentCustomer,
-  entity => isTruthy(entity) && !isTruthy(entity?.id),
+  (entity) => isTruthy(entity) && !isTruthy(entity?.id),
 );
 export const dataEntryCustomerQueries = {
   getCurrentCustomer,
@@ -36,4 +31,3 @@ export const dataEntryCustomerQueries = {
 };
 
 export const customerQueries = { ...odataGridCustomerQueries, ...dataEntryCustomerQueries };
-
