@@ -1,15 +1,13 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
-import { CERTIFICATIONS_FEATURE_KEY, State, CertificationsPartialState } from './certification.reducer';
+import { CERTIFICATIONS_FEATURE_KEY, State } from './certification.reducer';
 import { isTruthy } from 'imng-ngrx-utils';
 
-export const getCertificationsState = createFeatureSelector<CertificationsPartialState, State>(
-  CERTIFICATIONS_FEATURE_KEY,
-); 
+export const getCertificationsState = createFeatureSelector<State>(CERTIFICATIONS_FEATURE_KEY);
 
-const getCertifications = createSelector(getCertificationsState, state => state.gridData);
-const getLoading = createSelector(getCertificationsState, state => state.loading);
-const getGridODataState = createSelector(getCertificationsState, state => state.gridODataState);
-const getPagerSettings = createSelector(getCertificationsState, state => state.gridPagerSettings);
+const getCertifications = createSelector(getCertificationsState, (state) => state.gridData);
+const getLoading = createSelector(getCertificationsState, (state) => state.loading);
+const getGridODataState = createSelector(getCertificationsState, (state) => state.gridODataState);
+const getPagerSettings = createSelector(getCertificationsState, (state) => state.gridPagerSettings);
 export const odataGridCertificationQueries = {
   getCertifications,
   getLoading,
@@ -17,17 +15,14 @@ export const odataGridCertificationQueries = {
   getPagerSettings,
 };
 
-const getCurrentCertification = createSelector(
-  getCertificationsState,
-  state => state.currentCertification,
-);
+const getCurrentCertification = createSelector(getCertificationsState, (state) => state.currentCertification);
 const getIsEditCertificationActive = createSelector(
   getCurrentCertification,
-  entity => isTruthy(entity) && isTruthy(entity.id),
+  (entity) => isTruthy(entity) && isTruthy(entity.id),
 );
 const getIsNewCertificationActive = createSelector(
   getCurrentCertification,
-  entity => isTruthy(entity) && !isTruthy(entity.id),
+  (entity) => isTruthy(entity) && !isTruthy(entity.id),
 );
 export const dataEntryCertificationQueries = {
   getCurrentCertification,
@@ -36,4 +31,3 @@ export const dataEntryCertificationQueries = {
 };
 
 export const certificationQueries = { ...odataGridCertificationQueries, ...dataEntryCertificationQueries };
-
