@@ -1,15 +1,13 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
-import { EMPLOYEES_FEATURE_KEY, State, EmployeesPartialState } from './employee.reducer';
+import { EMPLOYEES_FEATURE_KEY, State } from './employee.reducer';
 import { isTruthy } from 'imng-ngrx-utils';
 
-export const getEmployeesState = createFeatureSelector<EmployeesPartialState, State>(
-  EMPLOYEES_FEATURE_KEY,
-); 
+export const getEmployeesState = createFeatureSelector<State>(EMPLOYEES_FEATURE_KEY);
 
-const getEmployees = createSelector(getEmployeesState, state => state.gridData);
-const getLoading = createSelector(getEmployeesState, state => state.loading);
-const getGridODataState = createSelector(getEmployeesState, state => state.gridODataState);
-const getPagerSettings = createSelector(getEmployeesState, state => state.gridPagerSettings);
+const getEmployees = createSelector(getEmployeesState, (state) => state.gridData);
+const getLoading = createSelector(getEmployeesState, (state) => state.loading);
+const getGridODataState = createSelector(getEmployeesState, (state) => state.gridODataState);
+const getPagerSettings = createSelector(getEmployeesState, (state) => state.gridPagerSettings);
 export const odataGridEmployeeQueries = {
   getEmployees,
   getLoading,
@@ -17,18 +15,9 @@ export const odataGridEmployeeQueries = {
   getPagerSettings,
 };
 
-const getCurrentEmployee = createSelector(
-  getEmployeesState,
-  state => state.currentEmployee,
-);
-const getIsEditEmployeeActive = createSelector(
-  getCurrentEmployee,
-  entity => isTruthy(entity) && isTruthy(entity.id),
-);
-const getIsNewEmployeeActive = createSelector(
-  getCurrentEmployee,
-  entity => isTruthy(entity) && !isTruthy(entity.id),
-);
+const getCurrentEmployee = createSelector(getEmployeesState, (state) => state.currentEmployee);
+const getIsEditEmployeeActive = createSelector(getCurrentEmployee, (entity) => isTruthy(entity) && isTruthy(entity.id));
+const getIsNewEmployeeActive = createSelector(getCurrentEmployee, (entity) => isTruthy(entity) && !isTruthy(entity.id));
 export const dataEntryEmployeeQueries = {
   getCurrentEmployee,
   getIsEditEmployeeActive,
@@ -36,4 +25,3 @@ export const dataEntryEmployeeQueries = {
 };
 
 export const employeeQueries = { ...odataGridEmployeeQueries, ...dataEntryEmployeeQueries };
-
